add tamago for u-root
All this involves is creating a tamago/main.go which is the existing main.go

Still a WIP, and do we need to move this to a new repo? Not sure.
The remaining issue is still runtime.Stack* variables which u-root has
to turn into Init functions.

the runtime.Stack* variables *could* be structured as leaf functions,
callable without using the stack on arm (since only LR is used in that case)
and risc-v (same -- uses LR for leaf functions)
(who cares about x86 any more anyway :-), which would nicely solve the problem.

Waiting for a comment from withsecure on how to solve the problem of early assembly
accessing variables which u-root has to rewrite as functions.

Signed-off-by: Ron Minnich <rminnich@gmail.com>